+++ requirements.txt
@@ -5,5 +5,6 @@ requests==2.28.1
 joblib==1.0.1
 scikit-learn==1.1.0
 pyodbc>=5.1.0
+Werkzeug==2.2.2
 pickle4
 gunicorn
