+++ utils/database_process.py
@@ -1,0 +1,22 @@
+# function for login checking
+def check_login(username, password, conn):
+    cursor = conn.cursor()
+    cursor.execute("SELECT * FROM Users WHERE Username = ? AND Password = ?", 
+                   (username, password))
+    result = cursor.fetchone()
+    return result is not None
+
+# get data for total predition
+def inputdata_to_totalpred(conn):
+    cursor = conn.cursor()
+    cursor.execute("select count(*) from Inputdata")
+    result = cursor.fetchone()[0]
+    return result
+
+# insert new input to database
+def insert_to_inputdata(data, conn):
+    cursor = conn.cursor()
+    cursor.execute("insert into InputData (Age, Sex, BMI, NumOfChildren, isSmoking," +
+                   "Region, Prediction, TimeDate) values(?, ?, ?, ?, ?, ?, ?, ?)", 
+                   (data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7]))
+    conn.commit()
